chore(api): hide generateClient from api/server subpath
