+++ packages/api/src/internals/index.ts
@@ -1,3 +1,4 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 export { InternalAPI, InternalAPIClass } from './InternalAPI';
+export { generateServerClient } from './generateServerClient';
